[REFAC] 차량 상세 뷰 분리 - 프로필뷰, 시세 범위 계산 뷰, 차량 상세 뷰로 분리
